refactor: отделить компоненты представления от контейнеров состояния
