+++ src/App.js
@@ -1,6 +1,6 @@
-import { PizzaConfigurator } from './components/PizzaConfigurator'
-import { PizzaSubmit } from './components/PizzaSubmit'
-import { PizzaResult } from './components/PizzaResult'
+import { PizzaConfigurator } from './containers/PizzaConfigurator'
+import { PizzaSubmit } from './containers/PizzaSubmit'
+import { PizzaResult } from './containers/PizzaResult'
 import { LABELS } from './values'
 
 function App() {
@@ -8,17 +8,8 @@ function App() {
         <div className="App">
             <header className="App-header">
                 <h1>{LABELS.configurator}</h1>
-                <h2>STAGING</h2>
                 <PizzaConfigurator />
-                <PizzaSubmit />
                 <PizzaResult />
-                <button
-                    onClick={() => {
-                        throw new Error('Some error')
-                    }}
-                >
-                    Break the world
-                </button>
             </header>
         </div>
     )
