+++ src/components/CheckboxGroup.js
@@ -1,27 +1,18 @@
-import { useSelector } from '../state'
-import { useDispatch } from '../PizzaProvider'
 import { LABELS } from '../values'
 import { LabeledPiece } from './LabeledPiece'
 
-export function PizzaChoice({ name, variants, selector, action }) {
-    const pizzaPropValue = useSelector(selector)
-    const dispatch = useDispatch()
-
-    const dispatchChange = (e) => {
-        dispatch(action(variants[e.target.value]))
-    }
-
+export const CheckboxGroup = ({ pizzaPropValue, dispatchChange, name, variants }) => {
     return (
         <LabeledPiece label={LABELS[name]}>
             {Object.entries(variants).map(([key, value]) => (
                 <div key={key}>
                     <input
-                        type="radio"
+                        type="checkbox"
                         id={key}
                         name={name}
                         value={key}
                         onChange={dispatchChange}
-                        checked={pizzaPropValue === value}
+                        checked={pizzaPropValue.includes(value)}
                     />
                     <label htmlFor={key}>{value}</label>
                 </div>
