+++ src/components/RadioGroup.js
@@ -1,29 +1,18 @@
-import { useSelector } from '../state'
-import { useDispatch } from '../PizzaProvider'
 import { LabeledPiece } from './LabeledPiece'
 import { LABELS } from '../values'
 
-export function PizzaMultiChoice({ name, variants, selector, actions: [add, remove] }) {
-    const pizzaPropValue = useSelector(selector)
-    const dispatch = useDispatch()
-
-    const dispatchChange = (e) => {
-        const payload = variants[e.target.value]
-        const action = (e.target.checked ? add : remove)(payload)
-        dispatch(action)
-    }
-
+export const RadioGroup = ({ pizzaPropValue, dispatchChange, name, variants }) => {
     return (
         <LabeledPiece label={LABELS[name]}>
             {Object.entries(variants).map(([key, value]) => (
                 <div key={key}>
                     <input
-                        type="checkbox"
+                        type="radio"
                         id={key}
                         name={name}
                         value={key}
                         onChange={dispatchChange}
-                        checked={pizzaPropValue.includes(value)}
+                        checked={pizzaPropValue === value}
                     />
                     <label htmlFor={key}>{value}</label>
                 </div>
